TL-24501 tool_totara_sync: explicitly close external db connections

+++ admin/tool/totara_sync/tests/user_external_database_test.php
@@ -236,18 +236,21 @@ class tool_totara_sync_user_external_database_testcase extends advanced_testcase
     }
 
     protected function tearDown() {
-        if ($this->configexists) {
-            // Drop sync table.
-            $dbman = $this->ext_dbconnection->get_manager();
-            $table = new xmldb_table($this->dbtable);
-            if ($dbman->table_exists($this->dbtable)) {
-                $dbman->drop_table($table, $this->dbtable);
+        if (isset($this->ext_dbconnection)) {
+            if ($this->configexists) {
+                // Drop sync table.
+                $dbman = $this->ext_dbconnection->get_manager();
+                $table = new xmldb_table($this->dbtable);
+                if ($dbman->table_exists($this->dbtable)) {
+                    $dbman->drop_table($table, $this->dbtable);
+                }
             }
+            $this->ext_dbconnection->dispose();
         }
+        $this->ext_dbconnection = null;
         $this->configdb = null;
         $this->config = null;
         $this->configexists = null;
-        $this->ext_dbconnection = null;
         $this->dbtype = null;
         $this->dbhost = null;
         $this->dbport = null;
